Don't trim whitespace for battery output
According to a 10 year old patch [0]:

> If the battery is not discharging it may be not possible to give
> information on > remaining time or consumption. The resulting strings
> (%remaining, %consumption, > %emptytime) are empty then. But because
> they are in the format string, the > output string contains unnecessary
> spaces in this case.

The code has changed in the meantime, but there is still a `trim` call
at the end of the print battery info method.

For my own setup, I like having one space of padding for every item in
the status bar. The lack of one for the battery section has been
annoying me for 2 years, so I decided to hack around it by simply
eliminating the offending `trim`.

- [0] https://gist.github.com/piotrdomagalski/3503688